Version 0.0.11
Changed the implementation of the grappling hook. It is now a indipendent node and no longer a child node of the playercontroller.
It was neccessary to do this because ohterwise the hooks move direction would be influenced by the movement of the player.

Added a hookable object
Added state pattern for the grappling hook (Idle, Shooting, Connected, Returning)
Hook can be shot to a hookable area and will connect. If connected the player can rotate around the connected area.

Problems:
- Dash works not correctly inside a gravity field -> open
- Teleport makes the player fly if he is in the idle sate after the teleport -> open