Bug Fix, can now leave Level again
because of some reason adding an Instantiated Scene as a Child of a Node this added scene is not properly considered as a child for some cases...
Because this goes way to in depth for my brain to handle i fixed the bug by changing the Leaving Function to manually deleting the currently played level. This isnt the best solution but i dont know what else to do.